adds basic navigation view, refactors contentview to gameview
